LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_timing_cross_origin_redirect_chain-expected.txt



PASS window.performance is defined
PASS window.performance.getEntriesByName is defined
PASS There should be one entry.
FAIL redirectStart == 0 in cross-origin redirect. assert_equals: redirectStart == 0 in cross-origin redirect. expected 0 but got 889
FAIL redirectEnd == 0 in cross-origin redirect. assert_equals: redirectEnd == 0 in cross-origin redirect. expected 0 but got 955.0000000000001
PASS fetchStart > 0 in cross-origin redirect.
FAIL startTime == fetchStart in cross-origin redirect. assert_equals: startTime == fetchStart in cross-origin redirect. expected 955.0000000000001 but got 889
